Project restructure, Blueprints clipped
